Added a frame handler that uses delta time instead of absolute time
This is often useful for games and other realtime applications.
