+++ src/main/kotlin/anim/Timelines.kt
@@ -1,41 +1,13 @@
 package kfx.anim
 
-import javafx.animation.Animation
 import javafx.animation.Interpolator
 import javafx.animation.KeyFrame
 import javafx.animation.KeyValue
-import javafx.animation.ParallelTransition
-import javafx.animation.PauseTransition
-import javafx.animation.SequentialTransition
 import javafx.animation.Timeline
 import javafx.beans.value.WritableValue
 import javafx.event.ActionEvent
 import javafx.event.EventHandler
 import javafx.util.Duration
-
-public inline fun <A : Animation> A.prep(play: Boolean = true, setup: A.() -> Unit): A {
-    setup()
-    if (play) play()
-    return this
-}
-
-public inline fun pause(
-    duration: Duration,
-    play: Boolean = true,
-    setup: PauseTransition.() -> Unit = {},
-): PauseTransition = PauseTransition(duration).prep(play, setup)
-
-public inline fun sequential(
-    vararg animations: Animation,
-    play: Boolean = true,
-    setup: SequentialTransition.() -> Unit = {},
-): SequentialTransition = SequentialTransition(*animations).prep(play, setup)
-
-public inline fun parallel(
-    vararg animations: Animation,
-    play: Boolean = true,
-    setup: ParallelTransition.() -> Unit = {},
-): ParallelTransition = ParallelTransition(*animations).prep(play, setup)
 
 public inline fun timeline(
     vararg keyFrames: KeyFrame,
